feat(product): 상품 가격을 원 단위 Long 으로 — price_won 확장·백필
Money(Long), 엔티티는 price_won 을 기준으로 읽고 옛 price 도 같은 값으로 쓴다(롤백 대비). 요청의 소수 원은
400(Jackson 이 Long 으로 받으면 조용히 자르므로 BigDecimal 로 받아 검사). 옛 컬럼 삭제는 다음 릴리스.

+++ product/feature/src/main/kotlin/com/kgd/product/application/product/usecase/CreateProductUseCase.kt
@@ -1,6 +1,5 @@
 package com.kgd.product.application.product.usecase
 
-import java.math.BigDecimal
 
 interface CreateProductUseCase {
     fun execute(command: Command, requester: ProductRequester): Result
@@ -13,7 +12,7 @@ interface CreateProductUseCase {
 
     data class Command(
         val name: String,
-        val price: BigDecimal,
+        val price: Long,
         val stock: Int,
         val brand: String? = null,
         val description: String? = null,
@@ -32,7 +31,7 @@ interface CreateProductUseCase {
     data class Result(
         val id: Long,
         val name: String,
-        val price: BigDecimal,
+        val price: Long,
         val stock: Int,
         val status: String,
         val sellerId: Long,
