+++ product/feature/src/main/kotlin/com/kgd/product/presentation/product/dto/ProductResponse.kt
@@ -3,12 +3,11 @@ package com.kgd.product.presentation.product.dto
 import com.kgd.product.application.product.usecase.CreateProductUseCase
 import com.kgd.product.application.product.usecase.GetProductUseCase
 import com.kgd.product.application.product.usecase.UpdateProductUseCase
-import java.math.BigDecimal
 
 data class ProductResponse(
     val id: Long,
     val name: String,
-    val price: BigDecimal,
+    val price: Long,
     val stock: Int,
     val status: String,
     val sellerId: Long,
